retrieval_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
faiss-cpu==1.9.0
numpy<2.0.0
protobuf<5
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
transformers>=4.40.0
peft>=0.17.0
accelerate>=0.30.0
safetensors>=0.4.0
tqdm>=4.66.0
requests>=2.31.0
# VTO Dependencies (IDM-VTON via Gradio Client)
gradio-client>=0.15.0
pillow>=10.0.0

